requirements.txt
RPLCD
pad4pi
RPi.GPIO
requests
configparser
